% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_42.R
\name{palette_42}
\alias{palette_42}
\title{Apply an ideas42 color scheme to your plots and figures}
\usage{
palette_42(name, n, type = c("discrete", "continuous"))
}
\value{

}
\description{
Apply an ideas42 color scheme to your plots and figures
}
\examples{

}
