% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_42_bright.R
\name{theme_42_bright}
\alias{theme_42_bright}
\title{Apply the ideas42 bright theme to your graphs}
\usage{
theme_42_bright()
}
\value{

}
\description{
Apply the ideas42 bright theme to your graphs
}
\examples{

}
