% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_hist.R
\name{viz_hist}
\alias{viz_hist}
\title{Generate a histogram with i42 styling}
\usage{
viz_hist(data, xvar)
}
\arguments{
\item{data}{data.frame}

\item{xvar}{variable}
}
\value{

}
\description{
Explore the distribution of your data with a histogram. This function generates
a ggplot2 object, so it's easily expanded.
}
\examples{

my_data <- data.frame(x = c(1,2,3,4,5))
viz_hist(my_data, x)
}
