% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_scatter.R
\name{viz_scatter}
\alias{viz_scatter}
\title{Generate a scatterplot}
\usage{
viz_scatter(data, xvar, yvar)
}
\arguments{
\item{data}{data.frame}

\item{xvar}{variable}

\item{yvar}{variable}
}
\value{

}
\description{
Explore the distribution of your data with a scatterplot This function generates
a ggplot2 object, so it's easily expanded with labels, themes, and other elements.
}
\examples{

x <- c(1,2,3,4,5)
y <-  rnorm(5, x, 0.5)
my_data <- data.frame(x = x, y = y)
viz_scatter(my_data, x, y)
}
